putting the markdown function in component_view and importing it where is it needed
